barbarfood-dev/requirements/local.txt
amqp==5.0.6
appdirs==1.4.4
apt-xapian-index==0.49
asgiref==3.4.1
astroid==2.7.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
billiard==3.6.4.0
black==19.10b0
blinker==1.4
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.2
celery==5.0.2
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.0.9
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
django-birthday==0.1.3
django-celery-beat==2.2.0
django-ckeditor==5.9.0
django-constance==2.7.0
django-cors-headers==3.2.1
django-debug-toolbar==2.1
django-extensions==2.2.5
django-filter==2.2.0
django-js-asset==1.2.2
django-phonenumber-field==4.0.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-rosetta==0.9.3
django-timezone-field==4.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
drf-writable-nested==0.5.4
drf-yasg==1.17.0
entrypoints==0.3
Faker==0.9.1
firebase-admin==5.0.1
flake8==3.7.9
google-api-core==1.31.2
google-api-python-client==2.19.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.0
google-crc32c==1.1.3
google-resumable-media==2.0.1
googleapis-common-protos==1.53.0
gpg===1.13.1-unknown
grpcio==1.39.0
httplib2==0.19.1
idna==2.8
inflection==0.5.1
ipdb==0.12.3
ipython==7.27.0
isodate==0.6.0
isort==5.9.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
keyring==18.0.1
kombu==5.1.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mixer==6.1.3
more-itertools==8.8.0
msgpack==1.0.2
mypy==0.761
mypy-extensions==0.4.3
netifaces==0.10.4
num2words==0.5.10
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.6.0
phonenumbers==8.11.1
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.3.0
pluggy==0.13.1
polib==1.1.1
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.17.3
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
PyExcelerate==0.8.0
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.10.2
pylint-django==2.0.13
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.3.0
pyotp==2.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pytest==5.3.2
pytest-cov==2.8.1
pytest-django==3.8.0
python-apt==2.0.0+ubuntu0.20.4.5
python-crontab==2.5.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.12.0
pytz==2021.1
PyYAML==5.3.1
redis==3.5.3
regex==2021.8.28
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sqlparse==0.4.1
systemd-python==234
text-unidecode==1.2
toml==0.10.2
traitlets==5.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
xmltodict==0.12.0
zeep==3.4.0
